:bug: :recycle: :sparkles: Reparado de SegmentationUI y modularizado para que otros componentes no tengan problemas, funcionalidad de guardado de recortes de espectros
